requirements.txt
intelhex>=2.3.0
pyserial>=3.5
progressbar2
